add a new utility app called: "Mock Silently"
Previously, "Mock my GPS" added the ability to use extras in geo-Intents
to silently control some of its runtime behavior.

The purpose for this utility is to intercept implicit geo-Intents
that originate from 3rd-party software,
and cannot be directly configured to include custom extras.

Once intercepted, custom extras are added to the geo-Intent.
Additionally, the geo-Intent is made explicit to open in "Mock my GPS".
Once the updated geo-Intent is started, the utility app closes..
without ever showing any visible UI.

Although it isn't a requirement,
it would be highly recommended to make the utility app the default to handle implicit geo-Intents.
This Android setting can always be revoked later.