add support for new extra: (boolean) "force_start"
When the "Mock my GPS" mocking Service is not running,
and an Intent is received which wishes to silently set a fixed position,
this flag determines whether to silently start the mocking Service and proceed.

The default behavior in this situation is to ignore the "silent_update" extra,
and display the app's UI to allow the user to manually start the mocking Service.

"Mock Silently" is also updated to support this new extra.